application/requirements.txt
flask
flask_sqlalchemy
psycopg2
loguru
names
python-ldap
requests
gevent